+++ tests/testthat/test-spiro_summary.R
@@ -11,6 +11,7 @@ test_that("stepwise summary works", {
 test_that("spiro_summary returns message when shortening interval", {
   expect_message(spiro_summary(gxt_data))
   expect_message(spiro_summary(ramp_data))
+  expect_snapshot(spiro_summary(ramp_data, interval = 300))
 })
 
 test_that("spiro_summary returns no message when set to quiet", {
@@ -21,4 +22,13 @@ test_that("spiro_summary returns no message when set to quiet", {
 test_that("spiro_summary excludes unfinished steps if desired", {
   expect_equal(nrow(spiro_summary(ramp_data, exclude = FALSE)), 25)
   expect_equal(nrow(spiro_summary(ramp_data, exclude = TRUE)), 24)
+})
+
+test_that("input is validated", {
+  expect_snapshot_error(spiro_summary(ramp_data, interval = -10))
+  expect_snapshot_error(spiro_summary(ramp_data, quiet = 1))
+  expect_snapshot_error(spiro_summary(ramp_data, exclude = "a"))
+  expect_snapshot_error(
+    spiro_summary(spiro(spiro_example("zan_gxt"), protocol = NA))
+  )
 })